аfix: fix delete func in 04-copy-directory

+++ 04-copy-directory/index.js
@@ -1,15 +1,22 @@
 const fs = require('fs').promises;
 const path = require('path');
-const constants = require('fs');
 
 const sourceDir = path.join(__dirname, 'files');
 const destinationDir = path.join(__dirname, 'files-copy');
 
 async function deleteDir(src) {
   try {
-    await fs.access(src, constants.F_OK);
-    // console.log("File found");
-    await fs.rm(src, { recursive: true });
+    const dirExists = await fs
+      .access(src, fs.constants.F_OK)
+      .then(() => true)
+      .catch(() => false);
+
+    if (dirExists) {
+      await fs.rm(src, { recursive: true });
+      console.log('Directory deleted successfully.');
+    } else {
+      console.log('Directory does not exist.');
+    }
   } catch (err) {
     console.error('Error deleting directory:', err.message);
   }
@@ -24,7 +31,14 @@ async function copyDir(sourceDir, destinationDir) {
       const sourcePath = path.join(sourceDir, item);
       const destinationPath = path.join(destinationDir, item);
 
-      await fs.copyFile(sourcePath, destinationPath);
+      const stat = await fs.stat(sourcePath);
+
+      if (stat.isFile()) {
+        const content = await fs.readFile(sourcePath, 'utf-8');
+        await fs.writeFile(destinationPath, content, 'utf-8');
+      } else if (stat.isDirectory()) {
+        await copyDir(sourcePath, destinationPath);
+      }
     }
 
     console.log('Directory updated successfully.');
